Introduces command line arguments with argparse
First argument is to locate the config file.

**Beware**: This patch might break previous installs, depending on how
the script was launched. However this change is necessary since we need
to get rid of `../config/` (a poor idea that ruins our flexibility).

The scripts now expects an explicit path to the config file, defaulting
to `config.yml` in the current directory.

Make sure to check `python collector.py -h` for the newly available help
message.